Moved admin table helpers
